Move util functions into binary module
